Add File/RoomPics For Room Management

+++ src/main/java/hotel/db/dto/room/RoomResponseDto.java
@@ -18,7 +18,7 @@ public class RoomResponseDto {
 	private String roomType;
 	private String bedType;
 	private Integer floorNumber;
-	private BigDecimal size;
+	private Double size;
 	private String roomDescription;
 	private BigDecimal price;
 	private String status;
